+++ docs/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## v7.3.0
 
+### 7.3.0 Features
+
+- `[Tree]` Added the ability to have separate icon button for expand/collapse and children count. ([#3847](https://github.com/infor-design/enterprise/issues/3847))
+
 ### 7.3.0 Fixes
 
 - `[Datagrid]` Fixed updatePagingInfo function which was wired on pager now and not working. To fix it i kept it flat on the datagrid object  `TJM` ([Issue #781](https://github.com/infor-design/enterprise-ng/issues/781))